+++ app/build.gradle
@@ -5,7 +5,7 @@ android {
     compileSdk 31
 
     defaultConfig {
-        applicationId "com.example.ltfrb_puvqrcodescanner"
+        applicationId "com.ltfrb.ltfrb_puvqrcodescanner"
         minSdk 22
         targetSdk 31
         versionCode 3
