+++ app/src/androidTest/java/com/ltfrb/ltfrb_puvqrcodescanner/ExampleInstrumentedTest.java
@@ -1,4 +1,4 @@
-package com.example.rama;
+package com.ltfrb.ltfrb_puvqrcodescanner;
 
 import android.content.Context;
