+++ app/src/main/java/com/ltfrb/ltfrb_puvqrcodescanner/ChangePassword.java
@@ -1,4 +1,4 @@
-package com.example.rama;
+package com.ltfrb.ltfrb_puvqrcodescanner;
 
 import android.app.ProgressDialog;
 import android.content.Intent;
