+++ app/src/main/java/com/ltfrb/ltfrb_puvqrcodescanner/LoginActivity.java
@@ -1,4 +1,4 @@
-package com.example.rama;
+package com.ltfrb.ltfrb_puvqrcodescanner;
 import androidx.appcompat.app.AppCompatActivity;
 import android.content.Intent;
 import android.os.Bundle;
@@ -6,12 +6,11 @@ import android.text.TextUtils;
 import android.view.View;
 import android.widget.Button;
 import android.widget.EditText;
-import android.widget.TextView;
 import android.widget.Toast;
-import com.example.rama.api.ApiClient;
-import com.example.rama.api.ApiInterface;
-import com.example.rama.model.login.Login;
-import com.example.rama.model.login.LoginData;
+import com.ltfrb.ltfrb_puvqrcodescanner.api.ApiClient;
+import com.ltfrb.ltfrb_puvqrcodescanner.api.ApiInterface;
+import com.ltfrb.ltfrb_puvqrcodescanner.model.login.Login;
+import com.ltfrb.ltfrb_puvqrcodescanner.model.login.LoginData;
 import retrofit2.Call;
 import retrofit2.Callback;
 import retrofit2.Response;
