+++ app/src/main/java/com/ltfrb/ltfrb_puvqrcodescanner/NetworkStateChecker.java
@@ -1,4 +1,4 @@
-package com.example.rama;
+package com.ltfrb.ltfrb_puvqrcodescanner;
 import android.annotation.SuppressLint;
 import android.content.BroadcastReceiver;
 import android.content.Context;
