+++ app/src/main/java/com/ltfrb/ltfrb_puvqrcodescanner/VolleySingleton.java
@@ -1,4 +1,4 @@
-package com.example.rama;
+package com.ltfrb.ltfrb_puvqrcodescanner;
 import android.content.Context;
 
 import com.android.volley.Request;
