+++ app/src/test/java/com/ltfrb/ltfrb_puvqrcodescanner/ExampleUnitTest.java
@@ -1,4 +1,4 @@
-package com.example.rama;
+package com.ltfrb.ltfrb_puvqrcodescanner;
 
 import org.junit.Test;
